+++ app/Models/UserModel.php
@@ -29,5 +29,6 @@ class UserModel extends ShieldUserModel
         'ruolo',
         'ultima_versione_vista',
         'richiede_cambio_auto',
+        'veicolo_id',
     ];
 }